requirements.txt
amqp==2.6.1
appdirs==1.4.4
asgiref==3.2.10
certifi==2020.6.20
chardet==3.0.4
distlib==0.3.1
dj-database-url==0.5.0
Django==3.0.3
django-crispy-forms==1.9.2
django-heroku==0.3.1
django-widget-tweaks==1.4.8
filelock==3.0.12
gunicorn==20.0.4
idna==2.10
Pillow==7.1.2
pipenv==2020.8.13
psycopg2==2.8.5
psycopg2-binary==2.8.5
pytz==2020.1
requests==2.24.0
six==1.15.0
sqlparse==0.3.1
urllib3==1.25.10
vine==1.3.0
virtualenv==20.0.31
virtualenv-clone==0.5.4
whitenoise==5.2.0
